+++ examples/examples.test.ts
@@ -1,4 +1,5 @@
-import { jsonParser } from "../examples/json";
+import { describe, expect, test } from "@jest/globals";
+import { jsonParser } from "./json";
 
 describe("JSON", () => {
     const jsonNumbers = ["0", "3.141592", "4.2195e1", "0.00E+00", "0e-0"];
@@ -13,15 +14,16 @@ describe("JSON", () => {
         ...jsonNumbers.map(jsonNumber => "-" + jsonNumber),
     ])("%o", json => {
         expect(jsonParser.parse(json)).toEqual({
-            pos: json.length,
-            val: JSON.parse(json),
+            success: true,
+            index: json.length,
+            value: JSON.parse(json),
         });
     });
     test("number parsing to fail..", () => {
-        expect(jsonParser.parse("00")).toBeNull();
-        expect(jsonParser.parse("- 0")).toBeNull();
-        expect(jsonParser.parse("0.")).toBeNull();
-        expect(jsonParser.parse(".0")).toBeNull();
+        expect(jsonParser.parse("00")).toHaveProperty("success", false);
+        expect(jsonParser.parse("- 0")).toHaveProperty("success", false);
+        expect(jsonParser.parse("0.")).toHaveProperty("success", false);
+        expect(jsonParser.parse(".0")).toHaveProperty("success", false);
     });
     test.each([
         ['"'],
@@ -35,6 +37,6 @@ describe("JSON", () => {
         ["u1234", "\u1234"],
     ])("escape %s", (escapeChar, char = escapeChar) => {
         const json = `"\\${escapeChar}"`;
-        expect(jsonParser.parse(json)).toEqual({ pos: json.length, val: char });
+        expect(jsonParser.parse(json)).toHaveProperty("value", char);
     });
 });