+++ src/primitive.ts
@@ -1,4 +1,6 @@
-import { Config, Source, pushError } from "./context";
+import { equals } from "emnorst";
+import type { Config, Source } from "./context";
+import * as error from "./error";
 import { Parser } from "./parser";
 import { updateState } from "./state";
 
@@ -6,11 +8,11 @@ import { updateState } from "./state";
  * Always succeed with the value of the argument.
  */
 export const pure = <T>(value: T): Parser<T> =>
-    new Parser(state => updateState(state, value, 0));
+    new Parser(state => updateState(state, value));
 
 export const fail = (): Parser<never> =>
     new Parser((state, context) => {
-        pushError(context, state.pos);
+        context.addError(error.unknown(state.i));
         return null;
     });
 
@@ -18,8 +20,8 @@ export const fail = (): Parser<never> =>
  * end of input
  */
 export const EOI = /* #__PURE__ */ new Parser((state, context) => {
-    if (state.pos < context.src.length) {
-        pushError(context, state.pos);
+    if (state.i < context.src.length) {
+        context.addError(error.unknown(state.i));
         return null;
     }
     return state;
@@ -32,14 +34,24 @@ export const EOI = /* #__PURE__ */ new Parser((state, context) => {
  * @example any.parse([]); // parse fail
  */
 export const ANY_EL = /* #__PURE__ */ new Parser((state, context) => {
-    if (state.pos < context.src.length) {
-        return updateState(state, context.src[state.pos], 1);
+    if (state.i < context.src.length) {
+        return updateState(state, context.src[state.i], 1);
     }
-    pushError(context, state.pos);
+    context.addError(error.unknown(state.i));
     return null;
 });
 
-export const el = <T>(value: T): Parser<T> => satisfy(srcEl => Object.is(srcEl, value));
+export const el = <T>(value: T): Parser<T> => satisfy(srcEl => equals(srcEl, value));
+
+export const oneOf = <T>(values: Iterable<T>): Parser<T> => {
+    const set = new Set<unknown>(values);
+    return satisfy(el => set.has(el));
+};
+
+export const noneOf = (values: Iterable<unknown>): Parser => {
+    const set = new Set(values);
+    return satisfy(el => !set.has(el));
+};
 
 export const satisfy = <T>(
     f:
@@ -49,26 +61,26 @@ export const satisfy = <T>(
     new Parser((state, context) => {
         let srcEl: unknown;
         if (
-            state.pos < context.src.length &&
-            f((srcEl = context.src[state.pos]), context.cfg)
+            state.i < context.src.length &&
+            f((srcEl = context.src[state.i]), context.cfg)
         ) {
             return updateState(state, srcEl, 1);
         }
-        pushError(context, state.pos);
+        context.addError(error.unknown(state.i));
         return null;
     });
 
 export const literal = <T extends Source>(chunk: T): Parser<T> =>
     new Parser((state, context) => {
-        if (state.pos + chunk.length > context.src.length) {
-            pushError(context, state.pos);
+        if (state.i + chunk.length > context.src.length) {
+            context.addError(error.unknown(state.i));
             return null;
         }
         for (let i = 0; i < chunk.length; i++) {
-            const srcEl = context.src[state.pos + i];
+            const srcEl = context.src[state.i + i];
             const chunkEl = chunk[i];
-            if (!Object.is(srcEl, chunkEl)) {
-                pushError(context, state.pos + i);
+            if (!equals(srcEl, chunkEl)) {
+                context.addError(error.unknown(state.i + i));
                 return null;
             }
         }