+++ src/string.ts
@@ -1,22 +1,18 @@
-import { pushError } from "./context";
+import * as error from "./error";
 import { Parser } from "./parser";
 import { updateState } from "./state";
 
-export interface RegExpGroupArray extends Array<string> {
-    groups?: RegExpExecArray["groups"];
-}
-
-export const regexGroup = (re: RegExp): Parser<RegExpGroupArray> => {
+export const regexGroup = (re: RegExp): Parser<RegExpExecArray> => {
     const fixedRegex = new RegExp(`^(?:${re.source})`, re.flags.replace("g", ""));
 
     return new Parser((state, context) => {
         if (typeof context.src !== "string") {
-            pushError(context, state.pos);
+            context.addError(error.unknown(state.i));
             return null;
         }
-        const matchResult = fixedRegex.exec(context.src.slice(state.pos));
+        const matchResult = fixedRegex.exec(context.src.slice(state.i));
         if (matchResult === null) {
-            pushError(context, state.pos);
+            context.addError(error.unknown(state.i));
             return null;
         }
         return updateState(state, matchResult, matchResult[0].length);
@@ -24,11 +20,13 @@ export const regexGroup = (re: RegExp): Parser<RegExpGroupArray> => {
 };
 
 export const regex: {
-    (re: RegExp, groupId?: never): Parser<string>;
+    (re: RegExp): Parser<string>;
     (re: RegExp, groupId: number | string): Parser<string | undefined>;
-} = (re, groupId = 0) =>
-    regexGroup(re).map(matchResult =>
-        typeof groupId === "number"
-            ? matchResult[groupId]
-            : matchResult.groups?.[groupId]!,
+    <T>(re: RegExp, groupId: number | string, defaultValue: T): Parser<string | T>;
+} = (re: RegExp, groupId: number | string = 0, defaultValue?: undefined) =>
+    regexGroup(re).map(
+        matchResult =>
+            (typeof groupId === "number"
+                ? matchResult[groupId]
+                : matchResult.groups?.[groupId]!) ?? defaultValue,
     );